\name{RPCLR-package}
\alias{RPCLR-package}
\alias{RPCLR}
\docType{package}
\title{
Implements the R-PCLR algorithm to calculate variable importance

}
\description{The R-PCLR algorithm can be used to estimate variable importance
in settings of high dimensional data arising from matched case-control studies. 
The algorithm accounts for the correlation between observations belonging to the same matched stratum, while incorporating some of the powerful features of Random Forests for evaluating the significance of high dimensional feature sets.

}
\details{
\tabular{ll}{
Package: \tab RPCLR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-19\cr
License: \tab GPL 2.0 \cr
LazyLoad: \tab yes\cr
}
}
\author{
Raji Balasubramanian

Maintainer: Raji Balasubramanian <rbalasub@schoolph.umass.edu>
}
\references{Balasubramanian, R., Houseman, E. A., Coull, B. A., Lev, M. H., Schwamm, L. H., Betensky, R. A. (2012). Variable importance in matched case-control studies in settings of high dimensional data, Submitted to Biostatistics. 
}
